refactord and added a single test suite
